requirements.txt
requests
pandas
numpy
streamlit
groq
python-dotenv
pathlib
